+++ src/main/java/META-INF/sun-jaxb.episode
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<bindings version="2.1" xmlns="http://java.sun.com/xml/ns/jaxb">
+  <!--
+
+Este arquivo foi gerado pela Arquitetura JavaTM para Implementação de Referência (JAXB) de Bind XML, v2.2.11 
+Consulte <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+Todas as modificações neste arquivo serão perdidas após a recompilação do esquema de origem. 
+Gerado em: 2017.01.11 às 06:43:19 PM BRT 
+
+  -->
+  <bindings scd="x-schema::tns" xmlns:tns="http://www.portalfiscal.inf.br/nfe">
+    <schemaBindings map="false">
+      <package name="br.inf.portalfiscal.nfe"/>
+    </schemaBindings>
+    <bindings scd="tns:distDFeInt">
+      <class ref="br.inf.portalfiscal.nfe.DistDFeInt"/>
+    </bindings>
+    <bindings scd="tns:retDistDFeInt">
+      <class ref="br.inf.portalfiscal.nfe.RetDistDFeInt"/>
+    </bindings>
+    <bindings scd="tns:resEvento">
+      <class ref="br.inf.portalfiscal.nfe.ResEvento"/>
+    </bindings>
+    <bindings scd="tns:resNFe">
+      <class ref="br.inf.portalfiscal.nfe.ResNFe"/>
+    </bindings>
+  </bindings>
+  <bindings scd="x-schema::tns" xmlns:tns="http://www.w3.org/2000/09/xmldsig#">
+    <schemaBindings map="false">
+      <package name="org.w3._2000._09.xmldsig_"/>
+    </schemaBindings>
+    <bindings scd="~tns:ReferenceType">
+      <class ref="org.w3._2000._09.xmldsig_.ReferenceType"/>
+    </bindings>
+    <bindings scd="~tns:SignedInfoType">
+      <class ref="org.w3._2000._09.xmldsig_.SignedInfoType"/>
+    </bindings>
+    <bindings scd="~tns:SignatureType">
+      <class ref="org.w3._2000._09.xmldsig_.SignatureType"/>
+    </bindings>
+    <bindings scd="~tns:SignatureValueType">
+      <class ref="org.w3._2000._09.xmldsig_.SignatureValueType"/>
+    </bindings>
+    <bindings scd="~tns:TransformsType">
+      <class ref="org.w3._2000._09.xmldsig_.TransformsType"/>
+    </bindings>
+    <bindings scd="~tns:TransformType">
+      <class ref="org.w3._2000._09.xmldsig_.TransformType"/>
+    </bindings>
+    <bindings scd="~tns:KeyInfoType">
+      <class ref="org.w3._2000._09.xmldsig_.KeyInfoType"/>
+    </bindings>
+    <bindings scd="~tns:X509DataType">
+      <class ref="org.w3._2000._09.xmldsig_.X509DataType"/>
+    </bindings>
+  </bindings>
+</bindings>
+
