--- src/main/java/com/fincatto/nfe310/classes/distribuicao/RetDistDFeIntObjectFactory.java
@@ -1,63 +1,0 @@
-//
-// Este arquivo foi gerado pela Arquitetura JavaTM para Implementação de Referência (JAXB) de Bind XML, v2.2.8-b130911.1802 
-// Consulte <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Todas as modificações neste arquivo serão perdidas após a recompilação do esquema de origem. 
-// Gerado em: 2017.01.06 às 03:24:59 PM BRT 
-//
-
-
-package com.fincatto.nfe310.classes.distribuicao;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the com.fincatto.classes.retDistDFeInt_v1_01 package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class RetDistDFeIntObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: com.fincatto.classes.retDistDFeInt_v1_01
-     * 
-     */
-    public RetDistDFeIntObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link RetDistDFeInt }
-     * 
-     */
-    public RetDistDFeInt createRetDistDFeInt() {
-        return new RetDistDFeInt();
-    }
-
-    /**
-     * Create an instance of {@link RetDistDFeInt.LoteDistDFeInt }
-     * 
-     */
-    public RetDistDFeInt.LoteDistDFeInt createRetDistDFeIntLoteDistDFeInt() {
-        return new RetDistDFeInt.LoteDistDFeInt();
-    }
-
-    /**
-     * Create an instance of {@link RetDistDFeInt.LoteDistDFeInt.DocZip }
-     * 
-     */
-    public RetDistDFeInt.LoteDistDFeInt.DocZip createRetDistDFeIntLoteDistDFeIntDocZip() {
-        return new RetDistDFeInt.LoteDistDFeInt.DocZip();
-    }
-
-}
