+++ src/main/java/org/w3/_2000/_09/xmldsig_/SignatureValueType.java
@@ -1,52 +1,49 @@
 //
-// Este arquivo foi gerado pela Arquitetura JavaTM para Implementação de Referência (JAXB) de Bind XML, v2.2.8-b130911.1802 
+// Este arquivo foi gerado pela Arquitetura JavaTM para Implementação de Referência (JAXB) de Bind XML, v2.2.11 
 // Consulte <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Todas as modificações neste arquivo serão perdidas após a recompilação do esquema de origem. 
-// Gerado em: 2017.01.06 às 03:24:55 PM BRT 
+// Gerado em: 2017.01.11 às 06:43:19 PM BRT 
 //
 
 
-package com.fincatto.nfe310.classes.distribuicao;
+package org.w3._2000._09.xmldsig_;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlID;
 import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
+import javax.xml.bind.annotation.XmlValue;
 import javax.xml.bind.annotation.adapters.CollapsedStringAdapter;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 
 /**
- * <p>Classe Java de KeyInfoType complex type.
+ * <p>Classe Java de SignatureValueType complex type.
  * 
  * <p>O seguinte fragmento do esquema especifica o conteúdo esperado contido dentro desta classe.
  * 
  * <pre>
- * &lt;complexType name="KeyInfoType">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="X509Data" type="{http://www.w3.org/2000/09/xmldsig#}X509DataType"/>
- *       &lt;/sequence>
- *       &lt;attribute name="Id" type="{http://www.w3.org/2001/XMLSchema}ID" />
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
+ * &lt;complexType name="SignatureValueType"&gt;
+ *   &lt;simpleContent&gt;
+ *     &lt;extension base="&lt;http://www.w3.org/2001/XMLSchema&gt;base64Binary"&gt;
+ *       &lt;attribute name="Id" type="{http://www.w3.org/2001/XMLSchema}ID" /&gt;
+ *     &lt;/extension&gt;
+ *   &lt;/simpleContent&gt;
+ * &lt;/complexType&gt;
  * </pre>
  * 
  * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "KeyInfoType", propOrder = {
-    "x509Data"
+@XmlType(name = "SignatureValueType", propOrder = {
+    "value"
 })
-public class KeyInfoType {
+public class SignatureValueType {
 
-    @XmlElement(name = "X509Data", required = true)
-    protected X509DataType x509Data;
+    @XmlValue
+    protected byte[] value;
     @XmlAttribute(name = "Id")
     @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
     @XmlID
@@ -54,27 +51,25 @@ public class KeyInfoType {
     protected String id;
 
     /**
-     * Obtém o valor da propriedade x509Data.
+     * Obtém o valor da propriedade value.
      * 
      * @return
      *     possible object is
-     *     {@link X509DataType }
-     *     
+     *     byte[]
      */
-    public X509DataType getX509Data() {
-        return x509Data;
+    public byte[] getValue() {
+        return value;
     }
 
     /**
-     * Define o valor da propriedade x509Data.
+     * Define o valor da propriedade value.
      * 
      * @param value
      *     allowed object is
-     *     {@link X509DataType }
-     *     
+     *     byte[]
      */
-    public void setX509Data(X509DataType value) {
-        this.x509Data = value;
+    public void setValue(byte[] value) {
+        this.value = value;
     }
 
     /**
